Factor tt::can_be_copy_constructed to its own file
